remove the example code from the repository.
considering that many issues were reported with lack of robustness of the example code, I have decided to remove it. while it was meant to initially be a demonstration of the bzip3 APIs, i do not want to turn it into a demonstration of how to use the C standard library. to guard the code against all sorts of errors (I/O, allocation) we would need to essentially triple the size of the example code for no reason in particular. one needs to check all I/O operations - fopen, fflush, fread, fwrite, fseek. you need to check malloc return values, need to check file headers, need to sync files using UNIX apis. then a handler for each case to display an error message and `exit`. as such, when the size of the "example" will grow 3 times and become just useless stuff completely irrelevant to someone who's using the library, it will stop serving its purpose.
the example code was not a part of the library, but to stop it from drawing so much attention, it has been removed.